// src/app/api/models/credencial-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Representação de Credencial do Usuário
 */
export interface CredencialDto {

  /**
   * Token de acesso
   */
  accessToken?: string;

  /**
   * Indica se o usuário está ativo
   */
  activeState?: boolean;

  /**
   * Email do Usário
   */
  email?: string;

  /**
   * Tempo de expiração do token de acesso
   */
  expiresIn?: number;

  /**
   * Id do Usuário
   */
  id?: number;

  /**
   * Login do Usuário
   */
  login?: string;

  /**
   * Nome do Usuário
   */
  name?: string;

  /**
   * Tempo de expiração do token de refresh
   */
  refreshExpiresIn?: number;

  /**
   * Token de refresh
   */
  refreshToken?: string;

  /**
   * Lista de permissões do Usuário
   */
  roles?: Array<string>;
}